@@ -1,22 +1,22 @@
 {
-  "name": "tech-world",
-  "version": "0.0.1",
+  "name": "driblet",
+  "version": "1.0.0",
   "private": true,
   "main": "index.jsx",
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
+    "@fontsource/roboto": "^5.1.1",
+    "@mui/icons-material": "^6.4.0",
+    "@mui/material": "^6.4.0",
     "@uiw/codemirror-extensions-basic-setup": "^4.23.7",
     "@uiw/codemirror-extensions-langs": "^4.23.7",
     "@uiw/codemirror-themes-all": "^4.23.7",
     "@uiw/react-codemirror": "^4.23.7",
     "dom-to-image-more": "^3.5.0",
-    "fontsource-roboto": "^4.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-material-ui-carousel": "^2.2.6",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3"
+    "react": "^19.0.0",
+    "react-color": "^2.19.3",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.1.2",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "install": "npm install react-scripts --save",
