@@ -1,6 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "fontsource-roboto";
+import { createRoot } from 'react-dom/client';
+import '@fontsource/roboto/300.css';
+import '@fontsource/roboto/400.css';
+import '@fontsource/roboto/500.css';
+import '@fontsource/roboto/700.css';
 import Application from "./application";
 
-ReactDOM.render(<Application id="application" />, document.getElementById("root"));
+
+const root = createRoot(document.getElementById('root'));
+root.render(<Application id="application" />);
